chore: move devtool-plugin under adapters
