+++ src/vue.ts
@@ -2,7 +2,7 @@ import Vue from 'vue'
 import { BG0, BM0, BA0 } from './core/base'
 import { Module, ModuleImpl } from './core/module'
 import { Store, StoreImpl, Subscriber } from './core/store'
-import { devtoolPlugin } from './devtool-plugin'
+import { devtoolPlugin } from './adapters/devtool-plugin'
 import { Dictionary, assert, bind } from './utils'
 
 let _Vue: typeof Vue
